+++ examples/resources/catalystcenter_assign_templates_to_tag/resource.tf
@@ -1,0 +1,4 @@
+resource "catalystcenter_assign_templates_to_tag" "example" {
+  tag_id       = "ea505070-6bb8-493f-bff0-8058e8e03ee5"
+  template_ids = ["75b0f85a-8157-4db3-ae2d-9807c893319a"]
+}