+++ examples/resources/catalystcenter_tag/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_tag.example "<name>"