+++ internal/provider/data_source_catalystcenter_tag_test.go
@@ -1,0 +1,69 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSource
+func TestAccDataSourceCcTag(t *testing.T) {
+	var checks []resource.TestCheckFunc
+	checks = append(checks, resource.TestCheckResourceAttr("data.catalystcenter_tag.test", "name", "Tag1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.catalystcenter_tag.test", "description", "Tag1 Description"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.catalystcenter_tag.test", "system_tag", "false"))
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDataSourceCcTagConfig(),
+				Check:  resource.ComposeTestCheckFunc(checks...),
+			},
+		},
+	})
+}
+
+// End of section. //template:end testAccDataSource
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
+// End of section. //template:end testPrerequisites
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceConfig
+func testAccDataSourceCcTagConfig() string {
+	config := `resource "catalystcenter_tag" "test" {` + "\n"
+	config += `	name = "Tag1"` + "\n"
+	config += `	description = "Tag1 Description"` + "\n"
+	config += `	system_tag = false` + "\n"
+	config += `}` + "\n"
+
+	config += `
+		data "catalystcenter_tag" "test" {
+			id = catalystcenter_tag.test.id
+			name = "Tag1"
+		}
+	`
+	return config
+}
+
+// End of section. //template:end testAccDataSourceConfig